HT17/generate_teachers/requirements.txt
amqp==5.0.1
asgiref==3.2.10
attrs==20.3.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.3.1
decorator==4.4.2
distro==1.5.0
Django==3.1.2
django-bootstrap-form==3.4
django-debug-toolbar==3.1.1
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-suit==0.2.25
djangorestframework==3.12.2
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
Faker==4.4.0
flake8==3.8.4
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
infinity==1.5
install==1.3.4
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
names==0.3.0
paramiko==2.7.2
phonenumbers==8.12.12
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
urllib3==1.25.10
validators==0.18.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
psycopg2-binary==2.8.*
gunicorn
python-memcached
